% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_memtrack.R
\name{start_memtrack}
\alias{start_memtrack}
\alias{stop_memtrack}
\title{FUNCTION_TITLE}
\usage{
start_memtrack(timestep = 1, trackmem = TRUE, trackCPU = FALSE,
  add_processes = "", outfile = tempfile(fileext = ".txt"))

stop_memtrack()
}
\arguments{
\item{timestep}{PARAM_DESCRIPTION, Default: 1}

\item{trackmem}{PARAM_DESCRIPTION, Default: TRUE}

\item{trackCPU}{PARAM_DESCRIPTION, Default: FALSE}

\item{add_processes}{PARAM_DESCRIPTION, Default: ''}

\item{outfile}{PARAM_DESCRIPTION, Default: tempfile(fileext = ".txt")}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2018) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}

Lorenzo Busetto, phD (2018) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
