% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_perftrack.R
\name{start_perftrack}
\alias{start_perftrack}
\title{start_perftrack}
\usage{
start_perftrack(timestep = 1, trackmem = TRUE, trackCPU = TRUE,
  trackdisk = TRUE, gc_onstart = TRUE, add_processes = "",
  outfile = tempfile(fileext = ".txt"))
}
\arguments{
\item{timestep}{\code{integer}, timestep used for recording performance, Default: 1}

\item{trackmem}{\code{logical}, If TRUE, monitor memory usage, Default: TRUE}

\item{trackCPU}{\code{logical}, If TRUE, monitor CPU usage, Default: TRUE}

\item{trackdisk}{\code{logical}, If TRUE, monitor disk usage, Default: TRUE}

\item{add_processes}{\code{character array} containg the names of other processes
to be monitored, Default: NULL}

\item{outfile}{\code{character}, Full path of a file used to store the csv file used to
store performance data. Defaults to a temporary file: tempfile(fileext = ".txt")}
}
\value{
The function is called for its side effects. It starts monitoring
performances and strore useful information in baseenv.
}
\description{
starts collection of information about memory, disk and CPU
usage of rsession and other processes
}
\details{
On Windows, the function calls \code{typeperf}. On Linux, it calls \code{???}
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2018) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
