% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_memtrack.R
\name{stop_perftrack}
\alias{stop_perftrack}
\title{FUNCTION_TITLE}
\usage{
stop_perftrack(gc_onexit = TRUE)
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\author{
Lorenzo Busetto, phD (2018) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
